refactor: move engine to a lib
