+++ programs/rules-engine/src/main.rs
@@ -1,44 +1,10 @@
-use chrono::{DateTime, Duration, Utc};
+use chrono::{Duration, Utc};
 use rand::Rng;
-use std::sync::mpsc::{self, Receiver};
+use std::sync::mpsc;
 use std::thread;
 use std::time::Duration as StdDuration;
+use rules_engine::{MonitoringEvent, RuleEngine};
 
-#[derive(Debug)]
-enum MonitoringEvent {
-    LastBatchTimestamp(DateTime<Utc>),
-    TxInclusionRate { address: String, rate: f64 },
-    TxReordered(String),
-}
-
-struct RuleEngine;
-
-impl RuleEngine {
-    fn handle_event(event: MonitoringEvent) {
-        match event {
-            MonitoringEvent::LastBatchTimestamp(ts) => {
-                let delay = (Utc::now() - ts).num_seconds();
-                if delay > 120 {
-                    println!("[!] Send Msg to Relayer: Sequencer delay detected: {}s since last batch.", delay);
-                }
-            }
-            MonitoringEvent::TxInclusionRate { address, rate } => {
-                if rate < 0.5 {
-                    println!("[!] Send Msg to Relayer: Possible censorship: {} inclusion rate = {:.2}", address, rate);
-                }
-            }
-            MonitoringEvent::TxReordered(tx_hash) => {
-                println!("[!] Send Msg to Relayer: Front-running suspected for tx: {}", tx_hash);
-            }
-        }
-    }
-
-    fn run(receiver: Receiver<MonitoringEvent>) {
-        for event in receiver.iter() {
-            RuleEngine::handle_event(event);
-        }
-    }
-}
 
 fn start_mock_data_feeder(sender: mpsc::Sender<MonitoringEvent>) {
     thread::spawn(move || loop {
